Refactored use case class for new pattern
