+++ application-module/domain-application-module/account-application-module/src/main/java/br/net/silva/business/usecase/CreateNewAccountByCpfUseCase.java
@@ -3,7 +3,7 @@ package br.net.silva.business.usecase;
 import br.net.silva.business.enums.TypeAccountMapperEnum;
 import br.net.silva.business.exception.AccountExistsForCPFInformatedException;
 import br.net.silva.business.mapper.MapToCreateNewAccountByCpfMapper;
-import br.net.silva.business.utils.ConverterUtils;
+import br.net.silva.daniel.utils.ConverterUtils;
 import br.net.silva.daniel.entity.Account;
 import br.net.silva.daniel.exception.GenericException;
 import br.net.silva.daniel.factory.CreateNewAccountByCpfFactory;
