+++ application-module/domain-application-module/account-application-module/src/main/java/br/net/silva/business/usecase/DeactivateAccountUseCase.java
@@ -2,7 +2,7 @@ package br.net.silva.business.usecase;
 
 import br.net.silva.business.enums.TypeAccountMapperEnum;
 import br.net.silva.business.mapper.MapToFindAccountMapper;
-import br.net.silva.business.utils.ConverterUtils;
+import br.net.silva.daniel.utils.ConverterUtils;
 import br.net.silva.daniel.entity.Account;
 import br.net.silva.daniel.exception.GenericException;
 import br.net.silva.daniel.interfaces.UseCase;
